+++ src/store/index.js
@@ -3,6 +3,7 @@ import Vuex from 'vuex'
 import user from './modules/user'
 import detail from './modules/detail'
 import cart from './modules/cart'
+import address from './modules/address'
 
 Vue.use(Vuex)
 
@@ -21,6 +22,7 @@ export default new Vuex.Store({
   modules: {
     user,
     detail,
-    cart
+    cart,
+    address
   }
 })