+++ src/store/modules/cart.js
@@ -39,7 +39,7 @@ export default {
     },
     async clearCartListAction (context, cartIds) {
       await deleteCartList(cartIds)
-      Toast('删除成功')
+      Toast.success('删除成功')
       // 再次调用请求商品数据的接口 重新更新渲染商品列表
       context.dispatch('getCartListAction')
     }