+++ src/utils/vant-ui.js
@@ -1,5 +1,7 @@
 import Vue from 'vue'
-import { Toast, Rate, Icon } from 'vant'
+import { Popup, Cascader, Toast, Rate, Icon } from 'vant'
+Vue.use(Popup)
+Vue.use(Cascader)
 Vue.use(Toast)
 Vue.use(Rate)
 Vue.use(Icon)